CMake: robust version extraction for P4Est
Make the version extraction in P4Est more robust: We should not rely on
the version string to define SUBMINOR and PATCH - this doesn't have to
be the case. So, always set P4EST_VERSION_SUBMINOR and
P4EST_VERSION_PATCH to "0" in FindP4EST.cmake if no valid number could
be extracted.