Update documentation a bit.
Reading through the documentation of some of the functions in DoFTools, there are a number of functions for which I can't quite figure out what exactly they do. This updates a couple of places, but more work remains. In particular, I have no idea from the description of what DoFTools::make_single_patch() does.