Patch by Alexander Grayver: Fix face interpolation between FE_Nedelec elements.
This patch fixes a bug in the FE_Nedelec::get_face_interpolation_matrix related to the wrong indexing of the DoFs located on element faces. This patch ensures that the code produces an interpolation matrix with full column rank meaning that all DoFs of the FE with lower order are constrained.