Testsuite: Remove 'Nightly' Track, add 'Continuous'
This commit removes the obsolete support for a 'Nightly' track and adds a
new 'Continuous' track. The idea is that in future 'tester' uses
exclusively the 'Regression Test' track and simserv04 uses 'Continuous'.

Both testing sites have different strategies: tester goes through all
commits with one configuration (for 32 and 64 bit indices), simserv04 just
checks the latest merge available but for a number of compilers and
configuration. The current mix of test results is just confusing and very
horrible to read :-/