Import those parts of BOOST 1.56.0 that we need for deal.II.
This includes renaming the existing directory and overwriting every
file with the corresponding one from BOOST 1.56.0. Then removing thos
parts of the boost sublibraries that we don't need but that for
some reasons were included previously. Specifically, this applies to
documentation and build directories in libs/ and to the fusion and
spirit libraries that consist exclusively of header files.