Added project_boundary_values_curl_conforming_l2() to VectorTools for use with FE_Nedelec elements in 2D & 3D.
This includes a test code included in tests/deal.II to check it works correctly.

This new function uses an L2-projection on edges and
faces (only edges in 2D, both for 3D) to compute
the constaints and factors in the residual from the edge
projection when computing the face constraints. The original
function project_boundary_values_curl_conforming() has
problems with non-rectangular faces, this one does not. The
function assumes that the face orientations are standard (e.g. a
sphere may cause problems due to mismatched face orientations).

The test shows that project_boundary_curl_conforming_l2() works on a
distorted cube (non-rectangular faces) for a simple test case
(polynomial function).