Implement rotation for vector valued dofs on periodic bc
This commit completes the interface for periodic boundary conditions. It
exposes the interpolation matrix (that is internally used to satisfy
hanging node constraints on periodic boundaries) and provides a possibility
to generate it from a rotation matrix and an index for the vector component
in question (in a similar vein as we already do this for different vector
valued finite elements).

Signed-off-by: Matthias Maier <tamiko@kyomu.43-1.org>