fix ParameterHandler::leave_subsection and parsing
- leave_subsection() had a bool return value for success that doesn't
make sense because an Assert() would never allow false to be returned.
Remove this.
- read_input() would not complain about unbalanced 'subsection'/'end' in
the input. Instead, the ParameterHandler would just stay in that
subsection and so weird error messages about undefined entries would
show up later.
- One could argue that read_input() should always start and end with an
empty subsection path, but we didn't enforce this in the past and it
might be useful to read separate files not into the "top level".