Add #include <iostream> to a few tutorial programs.
These are the tutorial programs that use std::cout or std::endl, which we need to
get from <iostream> but don't currently. Presumably because we include it transitively,
but for the sake of learning deal.II and C++, this is one more and unnecessary step
to understand. It turns out it also makes Eclipse happier if we do the explicit
#include, as it shows fewer source errors.