Remove a few 1D specializations for mapping.
The old implementation of MappingQEulerian only moved the vertices
in 1D, not interior points for higher order mappings. While it might
be debatable how useful non-linear mappings are in 1D, this was not
documented and lead to surprising behavior. By making the class more
general we can actually make our code base simpler because we can remove
a few template specialization and use the same code as in higher
dimensions.