Testsuite: Cleanup: Put "run" and "diff" stage into separate script
It was not a brilliant idea to begin with to exploit implementational (and
unstable) features of CMake's generators for our test logic. And the
breakage of the latest changes with the ninja generator evidence this fact.

Therefore, phase them out into its own sh script and just call it for the
test stage in mind.