Remove storing data we end up not using.
We passed a data element around and stored it in a local class, but
ended up not using it after all. This is likely a leftover from a
previous iteration in the code. In any case, removing the member
variable turns out to compile and test fine, so we don't need to
pass this data to the constructor either.