Simplify code.
There are a couple of places where we call a get_data() function
that returns a pointer to a base class when we know that the actual
data type is a pointer to derived class. We need to cast right back
to the derived class.

This can be avoided by using covariant return types where the get_data()
function returns a pointer to derived right away.