Testsuite: Bugfix, also pipe stderror in run_test.sh
The run_test.sh script supports two different output options for tests,
namely to directly write to a file 'output', or to just output on stdout.

This commit changes the behavior for the latter situation. It reroutes
stderr to stdout such that all output of a test is actually considered.

The former case turned out to be a problem with deallog that writes to
stderr.

Note: In case of a test writing to output, stderr is still lost.