Improve error messages about ghosted vectors.
The TrilinosWrappers::VectorBase class declared its own version of
ExcGhostsPresent that hid the class we already declare globally in
exceptions.h, and thereby led to a situation where we output less
information than is available. This also led to some confusion as
evidenced by a recent mail on the mailing list.

Fix this by removing the extraneous class. Also augment the error
message of the existing class.