Move a bunch of internal functions into anonymous namespaces.
The MappingQ1 class had a bunch of internal functions that were part of the class
declaration but are in fact only ever used internally and do not need a whole
lot of information from the class itself. Move these functions into an anonymous
namespace in the .cc file so as to keep the external interface of the class
as small as possible.