Bugfix: Fix return value of StoredIndex<1, S>
The problem is subtle. TensorAccessors::reordered_index_view returns an
internal object that behaves like a tensor.

The problem is that fully accessing the tensorial object, i.e., applying
operator[](unsigned int) rank times returned an object of type

  StoreIndex<0, [...]>

instead of type Number of the original tensor. The problem is that the
compiler has no way of knowing whether we want to cast this object to
Number during type matching, thus something like

  std::min( StoreIndex<0, [...]>(...), (double)4.);

fails.