Improve loop nesting.
We had a number of places of the form
  for (...tight loop...)
    if (constant condition)
      data update;

The compiler almost certainly can hoist the condition out of the loop,
but why make it this complicated for the compiler. I also find the code
easier to read because the loop is really only over a counting index
and does not carry any particular meaning at a level higher than the
if-statement.