fix expand_instantiation split logic
We optionally split instantiations in the generated .inst files based on
two macros SPLIT_INSTANTIATIONS_COUNT and SPLIT_INSTANTIATIONS_INDEX.
Intel 14 generates an error with an invalid % 0 in the old solution,
which happens if _COUNT is undefined. Fix this by defining a macro at
the beginning of the .inst file.