fix cmake 3.3 warning about code-gallery.h
cmake 3.3 will warn about CMP0058 and is unable to compile the documentation
when using Ninja:

 This project specifies custom command DEPENDS on files in the build tree
  that are not specified as the OUTPUT or BYPRODUCTS of any
  add_custom_command or add_custom_target:

   doc/doxygen/code-gallery/code-gallery.h

We have to jump through a few hoops to make this work: Generating a
dependency to a source file generated in a subdirectory CMakeLists.txt
requires a custom command attached to a custom_target. Because ADD_DEPENDENCIES
only allows custom_targets not custom_commands, we have to add another dummy.