Fix compilation.
GCC 4.6.3 does support C++11 but apparently runs into
trouble with the recent addition of the new Threads::new_task()
function. This patch ensures that the right overload is found
by not specifying a template argument explicitly, but explicitly
casting the argument to the type of the argument of the function
we want to call. This cast would have to be performed also when
specifying the template argument explicitly.