make expand_instantiations more robust
If the .inst.in files contains a mistake so that expand_instantiations
fails, the output will still be created and is then considered updated.
As a result, a second run of "make" happily continues without failing.
Fix this by creating a temporary file and only moving if the first call
succeededs.
Also make one error message more verbose.