Reset object to native state in IndexSet::clear().
IndexSet stores an index into an array. Upon construction, this index is set
to an invalid value (because the array is empty), whereas after IndexSet::clear()
the index was set to zero -- a perfectly reasonable, and consequently objectionable
index. This patch sets the index to an invalid value instead, as is also
done in the constructor.