Exit early if there are no matrix entries to add.
This gets around an issue where, if one is using a sparsity pattern
created by make_flux_sparsity_pattern and the 'nonzero' coupling option,
one may end up trying to add zeros to entries that do not exist in the
sparsity pattern on the current processor. This problem shows up when
adding fluxes between locally owned and ghost cells.

This commit sidesteps the issue by exiting early if there is
nothing (i.e., the only thing to do is add zero) to do.