Raise exceptions for invalid input.
subdivided_parallelepiped could generate cells with negative volume if
the provided edges were not ordered correctly. Try to help the user by
raising an exception if this is the case in subdivided_parallelepiped
instead of create_triangulation.

This commit also fixes tests that would fail as a result of removing the
orientation fix.

Partly reverts ccb56e7fb5.