Avoid use of std::is_default_constructible.
Some older compilers that have a -std=c++0x flag and that we deem
C++11 compatible do apparently not support std::is_default_constructible,
despite the fact that they support std::is_base_of. Therefore, avoid
the use of it. This does not matter here because just a few lines further
down, we call 'make_shared<T>()' which already requires default
constructibility, and so the user will get an error one way or the other.

While there, also ensure that we use the correct include file.