Use portable version of erfc().
erfc is only part of C99, and thus not in C++98. It did make it into C++ via the TR1
addition, which I think is in C++03 and consequently in C++11, but be that as it may,
let's just use the BOOST version of it to be on the safe side.