Get rid of the confusing ExcInvalidConstructorCall exception.
This exception did not print any useful error message, and it was used in a significant
number of places that just happened to check obscure conditions, without really
explaining in the error message what concretely they were doing. Replace all of these
with ExcMessage and provide explanations what exactly it is that's going wrong.