Disable an assertion in BOOST that GCC 6.x warns about.
GCC says that the compiler can assume that the address of the object that a reference
points to must be nonzero. That's correct. But the warning happens many times during
a build. Since this is more of a sanity check for what the BOOST people are doing
internally, I think there is little harm in just removing the assertion -- we
don't change the BOOST code anyway.