Update the bundled boost copy from 1.56 to 1.62.
This commit updates our bundled copy of boost from 1.56 to 1.62. While
doing this I removed several boost libraries that are not used anywhere
in deal.II nor in the rest of the part of boost that we bundle:

boost/compatibility: This little library wraps the standard C headers
and we don't need it.

boost/signals: This has long since been deprecated in favor of
signals2.

boost/tr1: The boost TR1 library contains reimplementations of features
that ultimately appeared in C++11. We implement this ourselves with the
std_cxx11 namespace so we skip this folder.

boost/thread/experimental: This library is new and we don't currently
have any use for it.