Remove SolverCG::res2.
This variable was only used inside the SolverCG::criterion() function that was
supposed to let derived classes overload when the solver terminates. But the
criterion() function was never called anywhere, so there is no point in
keeping it. Furthermore, the res2 variable it returns was never set. So
remove the whole shebang.