Fix the scopes of variables.
The current version has variables in outer scopes that are assigned the addresses
of local variables in inner scopes. These addresses are not actually used in the
outer scope after we leave the inner scope, but it's a bad idea anyway. Move
the declaration of the pointers to the inner scopes where they are actually
used.