Remove an unnecessary copy constructor.
PVS studio rightfully complains that we should define operator= whenever
we explicitly define a copy constructor. However, since our copy
constructor does nothing special it is in fact equivalent to the default
copy constructor, so we can just get get rid of it and let the compiler
generate the correct copying code.