Clean up GridIn::read_vtk() in several ways.
Rename some variables. Also note that we always had
  vertex_indices[i] = i
  quad_indices[i] = i
  line_indices[i] = i
and so can get rid of these maps. They are a left-over from apparently copy-pasting
the initial version of this function from the UNV reader where this did not
have to be the case.

The function has many other problems, but I'm an incrementalist and I choose progress
over waiting for the perfect to happen :-).