Do not instantiate invalid classes.
We have numerous places where we do something of the sort

  template <int dim, int spacedim>
  void foo (const Triangulation<dim,spacedim> &triangulation) {
    if (dim == 3)
      {
        typename Triangulation<3,spacedim>::cell_iterator cell = triangulation.begin();
        ...

Since we only get into this piece of code if dim==3, there is nothing wrong with
this: if dim==3, then spacedim>=3. On the other hand, the compiler will still
instantiate Triangulation<3,spacedim> even if dim==spacedim==1.

This patch works around this by replacing the type by
  Triangulation<3,max(3,spacedim)>
which for all of the cases in question leads to the exact same type, but
avoids instantiating invalid types.