Redo output for physics/step-44* tests
Instead of checking the convergence rates for each Newton iteration
against some benchmark result that is machine specific, we now monitor
the solution at a fixed point. This should remain unchanged if the
problem converges correctly for each timestep. The nonlinear solver
output is now send to the file "stdout".

Fixes #3747