Fix the polynomial degree the ABF element reports about itself.
As stated in the original paper (http://www-users.math.umn.edu/~arnold/papers/vecquad.pdf),
section 5, and as also stated in our discussion of the PolynomialsABF element, the
ABF space of order 'r' actually contains polynomials of degree 'r+2'. Report this
accurately.

Without this, the computation of embedding matrices fails in 3d because we integrate
the least squares matrix terms with a quadrature formula of too low order.