Extensions for Trilinos support in LinearOperators.
Trilinos sparse matrices and preconditions can now we wrapped as
LinearOperators, and can be used (nearly) as naturally as the deal.II
linear algebra classes could. This means that they can be used as a
transpose_operator(), inverse_operator() and within the
schur_complement().

The core support for block operations is also there, but has not been
thoroughly tested.