Refactor SchurComplement operator.
Subsequent to the implementation of Payloads for LinearOperators, it is
necessary that compound operations such as this one be expressed in
terms of overloaded operations (operator+,* etc.) instead of in a
low-level format, with the operator functions v_mult etc. expanded
manually. This is because the payload is not correctly configured in the
latter case, and the operations defined on the Payload-level would not
align with those being performed by the higher-level LinearOperator
interface. When operator+,* etc. are called, the equivalent operations
for the payload are automatically set up.