Fix, and deprecate, a PETSc parallel vector copy constructor.
The former implementation of this copy constructor contained a subtle bug: since
VectorBase::operator= is implicitly defined, calling it simply copies (byte for
byte) the members of a VectorBase. This is erroneous because now two different
destructors will call VecDestroy on the underlying Vec.

The fix is simple: (nearly) preserve the current behavior by just deep copying
the vector.

Since one should never have an explicit instance of VectorBase (and there is no
reasonable way to copy a parallel vector and then enforce a possibly different
local_size) this constructor I also marked this constructor for deprecation: one
should use one of the concrete classes (Vector or MPI::Vector) in applications.