Tests: Remove build_tests/ category
The build_tests/ category had been an ugly hack for compiling all
configurable example steps.

But nowadays we already compile all (configurable) example steps as part
of the main build process - so this is redundant. Further, selectively
running a subset of example steps (as was done) is ugly and useless
(output is never compared).

Thus, let's simply remove the category and custom cmake script files.