Add a .clang-format file
This commit adds a .clang-format [1] file that approximates our indent
style as defined by
  ./contrib/styles/astyle.rc

The motivation behind this commit is the simple fact that clang-format
(utilizing the clang lexer and tokenizer) is a far superior indenter
than any of the alternatives (including astyle).

The format file tries to give a close approximation to the output of
astyle. Unfortunately astyle misinterprets heavily templated stuff
quite a bit, so that
  astyle after clang-format  =  clang-format
does not always hold.

[1] https://clang.llvm.org/docs/ClangFormatStyleOptions.html