Use only enum values, don't look at bits.
This fixes an (undocumented) oddity in ParameterHandler::print_parameters(): We
were at times looking at individual bits instead of just the declared values
of ParameterHandler::OutputStyle. This presumably allowed for calling that
function with a combination of the OutputStyle flags, for reasons that no
longer seem particularly relevant nor obvious.

This patch removes this possibility from the current implementation of
the function, but retains it for the (deprecated) function
ParameterHandler::print_parameters_section().