Compile TransfiniteInterpolationManifold<1> in the library.
In addition, disable usage of the class by throwing an exception.

This class is not as useful in 1D as it is in higher spatial dimensions, but
compiling it in the library improves generic programming (i.e., one can write a
solver that runs in 1D and 2D and on a 2D execution path can use
TransfiniteInterpolationManifold).