Be careful for which cells and objects we allocate hp DoF indices.
The patch itself is actually pretty small -- we just have to occasionally exclude cells
when we loop over them, specifically of they are artificial. The patch, however, is
far larger because we replicated the same code for cells, faces, and vertices for
the 1d, 2d, and 3d implementations. I didn't feel like making the same changes
3 times over, so I refactored these things into functions that deal with cells,
faces, and vertices. This makes the overall patch rather large.