Overhaul GridTools::internal::fix_up_object.
The primary issue with excising Boundary from the library is the lack of
an equivalent project_to_surface function in the Manifold code (since
Manifolds don't know anything about faces or lines, just points and
geodesics). Fortunately, this function is only called in one place, and
the default implementation in StraightManifold is the only
implementation in the library: hence, we get around this issue by
copying and pasting StraightManifold::project_to_surface into the new
internal project_to_object function and marking this change as
incompatible.

To the best of the author's knowledge, no users have implemented their
own custom project_to_surface nor are they using the one function that
calls this (GridTools::fix_up_distorted_child_cells) so this small
incompatibility is acceptable.