Make DoFAccessor::set_active_fe_index() a 'const' function.
We are not being very consistent in marking which DoFAccessor functions are 'const'
because it's difficult to actually define what 'const' would mean: does it change
the accessor, or does it change the DoFHandler it points to. As a consequence,
the TriaAccessor member functions are *all* const, but the DoFAccessor functions
are more or less random.

Rather than attacking this problem altogether, I'm making the minimal change that
solves the problem I currently have.