Fixes possible error on "Vector<Number>::add(..)".
In "Vector<Number>::add(..)", a pointer to the first element of the
std::vectors was gotten, for instance, by &indices[0], which implies
that the first element of the vector exists. Therefore, if an empty
vector was used, while this works on gcc, this would result in a crash
when using Visual Studio. In this case, the pointer should be gotten by
std::vector::data(). This fix changes de former by the latter access form.