Improve tolerances for errors near zero.
Details in the floating point implementation (e.g., whether or not -ffast-math
is used) can change the error by an order of magnitude or more. Get around this
by, in each relevant case, bounding the error above.

Since the errors are now checked in a different way the alternative output file
has been removed.