Avoid where possible to use std::ostringstream to generate filenames.
To students, this seems like an unnecessary detour. This may have been useful in
olden times, but we now have helper functions to convert numbers to text without
having to create a std::ostringstream in user code, and we also have std::to_string().

Finally, this avoids the awkward idiom
  std::ostringstream filename;
  ...
  std::ofstream output (filename.str().c_str());
where we have to call .str().c_str().

Instead of this, just create the filename in place and be done with it.