Convert an Assert into static_assert.
The TableIndices class had Assert statements in a couple of places
that the number of indices in question is a positive number, not zero.
Proximally, this led to Coverity complaining that a function called in a
noexcept function of Table may in fact throw an exception. The warning
is correct, and is most easily worked around by converting the
assertion into a static_assert -- which has to hold for the entire
class, so that's where I moved it from the affected member functions.