Fix iterating over entries of parallel PETSc matrices.
Previously, it was not possible to iterate over the local range because
one would have to call
  matrix.end(row)
where 'row' is the last locally owned row, and that triggered an
assertion because this end iterator is also the begin iterator of
the next row -- which is not locally owned any more.

Fix this.