Tests: Do not print memory consumptions in fe/fe_move_0? tests
The memory consumption is simply not stable with respect to architecture
and standard library.

The test remains useful - the actual contents of the moved elements is
still printed and compared in a unique (and stable) manner.

Further, remove output variants for clang9_darwin - I checked that the
output is now the same.