Improve alignment of \s in macros.
This is a bit tricky: doxygen wants to place the first '{' all the way to the
left regardless of its indentation (and indent the rest of the lines in each
macro normally), but astyle does not like left-justified macro blocks. Resolve
the issue by telling astyle to leave preprocessor blocks alone and
left-justifying all preprocessor blocks.