Avoid writing colinear points in GridOut::write_gnuplot.
GridOut::write_gnuplot has an option where, for plotting curved grids, it can
plot multiple line segments per face to give the illusion of a cell with curved
faces. This procedure creates massive output files since every face of every
corresponds to many individual line segments.

This commit implements a compression scheme where we do not output intermediate
points that are colinear. This lowers the number of points in a polar grid
output file by about 50% (i.e., the radial cell faces are not curved and can be
plotted with two points instead of perhaps five or six; same for the interior
cell faces).