Update the GNUPlot flags.
This commit fixes two intertwined issues with the current set of flags:
1. The variable names and descriptions are written in a way that implies
only boundary lines can be curved (which is no longer true). This commit
renames n_boundary_face_points to n_curved_line_points to make this
clearer.
2. write_gnuplot currently writes a lot of extra lines for boundary
faces in 3D and there is no way to turn this off. I would like to use
this function to generate wireframes of meshes, so I do not want this
extra output. I added a boolean (defaults to true) that controls whether
or not these extra lines are plotted.