Extend template specialisations for intermediate AD number types.
Its not likely, but it might be possible that in some complex chain of
operations we're left with one of these intermediates that needs to be
operated on or used. If this is the case then having these
specialisations facilitate the driver and helper classes to function
with these number types.