Execute quick tests in parallel.
The quick tests are currently set up as a custom target which immediately
launches ctest. The difficulty with this is that its not possible (AFAICT) to
get the number of jobs available from make at that point, so there is no way to
run these in parallel without giving an explicit number of jobs to ctest.

This PR sets the number of jobs to the number of CPUs on the machine: while
hardcoding this is not ideal, this is the right usage pattern for a build script
like

    make -j20
    make test

where all processors on the machine should be available.