Rework implementation of taping within ADHelperBase.
This commit abstracts the notion of taping almost completely away from
the ADHelperBase class. The specifics of the taping process are now
fully contained within the driver classes (at the expense of having to
have some non-static interfaces to the driver classes). However, this
should better enable further extension to use other taped AD
implementations in the future.