Fix bugs in extract_boundary_mesh.
In 3d, extract_boundary_mesh swaps vertices to get consistent boundary cell orientations. However, this swapping was so far not taken into account during the assembly of the map between boundary cells and corresponding faces of the volume cells. Fixing the latter required substantial changes (a quick and dirty fix would have been possible in principle, but certainly not clean) to extract_boundary_mesh because the (local) face indices of the volume cells must be known when the map is assembled. Also, the algorithm to do the refinement of the boundary mesh has been modified such that map lookups are no longer necessary. Moreover, vertex swapping has been introduced also for the 2d case in order to make the 2d behavior consistent with the 3d behavior (in previous versions of the function, the normal vector of the boundary mesh pointed always into the domain in 3d, whereas it pointed either into or out of the domain in 2d depending on the situation). Further, the normals have been adjusted to generally face outward.  Fixes #7467.