CMake: Bugfix: Guard DEAL_II_HAVE_OPENMP_SIMD test
We also have to guard the DEAL_II_HAVE_OPENMP_SIMD test with
DEAL_II_ALLOW_PLATFORM_INTROSPECTION. If the latter is set to false our
cmake configuration does not attempt to determine any specifics of the
unerlying CPU or architecture (and instead relies on the user specifying
everything by hand). This is necessary for scenarios where deal.II gets
compiled on a different machine than the one it will be run on
eventually.

In reference to #7663.