Cleanup the Vector::thread_loop_partitioner logic.
It was previously possible to, in one function call, either set up the
partitioner multiple times or for inappropriately small vectors. This
commit cleans up the way we handle the partitioner in multiple places so
that, when possible, vectors share partitioners and do not set up
partitioners if they are too small.