[CI]: add Jenkins.mark
Add a separate Jenkins pipeline that runs in a few seconds to set the
commit status to pending. This is needed as Jenkins only starts setting
a commit status when the normal jobs start running and not when they are
in the queue. This can lead to a) accidental merges before we test
things and b) repeated /rebuild triggers.