Run setting canonical pages as part of the doxygen generation process.
This has the advantage that we need not worry about when that POST_BUILD
command is run. In particular, there can not be any race conditions.
This is, however, only true for the doxygen generated pages, whereas the
script also touches the other .html files for which the target I put the
script under has no in- or out-dependencies.