release: wrap_comment warnings
Not all of them here might be crictical, but some are clearly bugs
(@name). In general, if @ref is not in a single line, doxygen sometimes
(not always) fails to create the correct link.
The line breaks in differentation/ are likely going to fix some of the
doxygen complaints about missing links.