examples: ignore temporary cc files in cmake
I had an editor leave a temporary cc file in the example directory. This
caused cmake/make to fail in hilarious ways. To try it yourself:
``touch examples/step-63/a#.cc``

By reducing the glob to files that are of the form "step-*.cc", we
should be able to avoid this problem. In my case the file was
".#step-63.cc".