Add std::complex support to integrate_difference()
The type of exact_solution was double and fe_function could have any numerical
type (double, float, std::complex<double> or std::complex<float>). This would
lead in certain cases to unnecessary casts, for example from double to float.
In addition it was not possible to compare a std::complex exact_solution to a
std::complex fe_function.

Now the types of exact_solution and fe_function must be the same, as a result
integrate_difference() can be used to compare a std::complex exact_solution to
a std::complex fe_function.

This change is not backward-compatible. The old version of the function has
been deprecated.