Extract dofs per component
Extract the set of local, active or relevant DoF indices for each

components within the mask that are owned by the current  processor. For
components disabled by the mask, an empty IndexSet is returned.

For regular DoFHandler objects, this set is the complete set with all
DoF indices. In either case, it equals what
DoFHandler::locally_owned_dofs() returns. The ownership type is fixed by
a small enum that controls which type of ownership is asked off.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>