Added the colorize option to subdivided_hyper_cube grid generator
The subdivided_hyper_cube is the one of the only grid generator which
does not allow the boundaries to be colorized. This is a non-standard
behavior, because all of the other generators (hyper_cube,
hyper_rectangle, subdivided_hyper_rectangle) allow one to colorize the
boundaries. This small pull request addresses this by adding the
colorize option to the subdivided_hyper_cube grid. It also fixes the
grid_generator_from_name to make it coherent with the new
implementaiton.