Fix 1D bug in one version of make_flux_sparsity_pattern.
When a neighbor has children, the function loops over each face and the
children of each face_iterator. But in 1D the number of children of a
face is always zero. The result is that flux-couplings are missed in
the 1D case when not all cells are on the same level.

Fix this by avoiding to iterate over the children of the face in 1D.
This is the same solutions that is already used in line 597 of
source/dofs/dof_tools_sparsity.cc.