Add header quadrature_generator.h
Add a header to be used for implementing a QuadratureGenerator class, which
should generate immersed quadrature rules for the different regions of a cell
intersected by the zero contour of a level set function. Add a hypercube class
in a internal namespace QuadratureGeneratorImplementation, which is needed to
implement the underlying algorithm.