Test vector_tools/integrate_difference_04_hp_02: Make test less expensive
The test runs in 3D in debug mode on a 3 times globally refined mesh.
Depending on workload of the machine this test is running on this might
imply to hit the 10 minute timeout. Make the test 8 times less expensive
by reducing the number of global refinement steps

In reference to #10073
See also: https://cdash.43-1.org/viewTest.php?onlydelta&buildid=6260