do not print mesh in opencascade/stl_* tests
These tests create a mesh (with our opencascade) interface.
Unfortunately, the result pretty much depends on the mesher being used
(which changes between opencascade and oce versions), the standard
library version (libstdc++ versus libc++) and the CPU architecture.

I initially thought that the test result would not depend on the machine
this test is running on and created the output alternatives
(*.output.opencascade7.4.0) on my laptop. Evidently this did not fix
this test on our main tester. Let's simply remove the output and just
check that the code path runs without throwing an error.