Add classes restricting a function to a lower dimension
Add a class CoordinateRestriction that corresponds to restricting one
coordinate of a dealii::Function, f, to a given value. That is, if Z is
a fixed value we create the new function

g = g(x, y) = f(x, y, Z).

Correspondingly, add a class PointRestriction that takes a point (Y, Z)
and restricts a dealii::Function to the point, i.e, we create the new
function

g = g(x) = f(x, Y, Z).