Fix the number of entries in the FEValues view cache.
Looking at the implementation of the FEValuesViews::Vector and similar classes, we
see that they return 'spacedim' objects at a time, and similar for the tensor views.
The number of entries in the cache for vectors was already keyed off 'spacedim',
but for the tensors it was based on 'dim' instead. This isn't quite right.