Use range-based for loops.
These places also had some kind of complicated logic with only considering
'used' cells, but that is equivalent to using a non-raw iterator. Furthermore,
we only cared about cells that have the refine flag set, and that can only
happen if the cell is active. So just loop over the active cells instead.