Fix compilation with intel 17 (and a UCD bug).
In particular, the new function set_node_numbers sets up the internal node
numbers for patches in the deal.II numbering, not UCD or DX or anything else.
However, the code

    nodes[GeometryInfo<dim>::ucd_to_deal[0]] = start;

assumes the RHS values are numbered with UCD (e.g., that start is the global
index for local UCD vertex number 0), which is not correct. Hence: fixing this
bug corrected the UCD output for many tests.