Unify internal names of SUNDIALS IDA interfaces.
As previously done for the KINSOL interfaces, this patch avoids the t_dae_ prefix
and instead uses _callback as suffix. I don't quite know where the t_ prefix
comes from, but it sounds like jargon that may be present in the SUNDIALS
internal documentation but that a typical reader of our code would not
know about.

Since these are all internal functions of deal.II, we can rename them freely.
In particular, because they are in anonymous namespaces, we can re-use the same
names already in use in the KINSOL interfaces.