Enable get_face_interpolation_matrix() in 1D on FE_Q and FE_Bernstein
Currently, get_face_interpolation_matrix() on FE_Q and FE_Bernstein
throws an ExcImpossibleInDim(1) exception in 1D. This is strange
because the interpolation matrix should just be a 1-by-1 unit matrix.
Remove the dim > 1 assert in FE_Q_base and FE_Bernstein to fix this
and modify two tests to check that it works.