TBB oneAPI: make global_control object static
tbb::global_control is a class that affects the specified behavior of
tbb during its lifetime. Thus, in order to set a global thread limit for
tbb we have to maintain the object throughout the execution of the
program. We do this by maintaining a static std::unique_ptr.

A std::unique_ptr is a good choice here because tbb::global_control does
not provide a mechanism to override its setting - we can only delete the
old and replace it with a new one.