base/polynomials_bdm.cc: avoid a compiler warning with gcc-12
gcc-12 in release mode will inline the evaluate function into some
methods of the FE_BDM class. It then rightfully complains that the (now
modified) else branch will write out of bounds.

Despite the fact that we instantiate this polynomials class and the
related FE class for 1d this code has never worked. Thus, fail
prominently with an error message.