tests cgal/* add output variants
The following tests print a cgal surface mesh for comparison to the
output file:

  cgal/cgal_surface_mesh_02
  cgal/cgal_surface_triangulation_03
  cgal/cgal_triangulation_06

Unfortunately, the output doesn't seem to be stable. For example,
cgal/cgal_surface_triangulation_03 fails for cgal-5.2.1 and succeeds for
cgal-5.4.1. I fear that the precise ordering and output of the meshes is
not stable and depends on the userland and what library versions are
used.

Let us simply not print the final result but just check that CGAL
succeeded in creating a corresponding mesh.